refactor: moved cdf.calls into root of utils module at utils.retrieve
